+++ eth_defi/enzyme/deployment.py
@@ -174,6 +174,25 @@ class EnzymeDeployment:
         tx_hash = value_interpreter.functions.addPrimitives(primitives, aggregators, rate_assets).transact({"from": self.deployer})
         return tx_hash
 
+    def remove_primitive(
+        self,
+        token: Contract,
+    ) -> str:
+        """Remove a primitive asset to a Enzyme protocol.
+
+        This will tell Enzyme how to value this asset.
+
+        - See ChainlinkPriceFeedMixin.sol
+
+        :return:
+            Transaction hash for the addition
+        """
+        assert isinstance(token, Contract), f"Got bad token: {token}"
+        value_interpreter = self.contracts.value_interpreter
+        primitives = [token.address]
+        tx_hash = value_interpreter.functions.removePrimitives(primitives).transact({"from": self.deployer})
+        return tx_hash
+
     def create_new_vault(
         self,
         owner: HexAddress,
@@ -183,6 +202,7 @@ class EnzymeDeployment:
         shares_action_time_lock: int = 0,
         fee_manager_config_data=b"",
         policy_manager_config_data=b"",
+        deployer=None,
     ) -> Tuple[Contract, Contract]:
         """
         Creates a new fund (vault).
@@ -195,6 +215,11 @@ class EnzymeDeployment:
             Tuple (Comptroller contract, vault contract)
         """
 
+        if not deployer:
+            deployer = self.deployer
+
+        assert deployer, "No deployer account set up"
+
         fund_deployer = self.contracts.fund_deployer
         tx_hash = fund_deployer.functions.createNewFund(
             owner,
@@ -206,7 +231,7 @@ class EnzymeDeployment:
             policy_manager_config_data,
         ).transact(
             {
-                "from": self.deployer,
+                "from": deployer,
             }
         )
         receipt = self.web3.eth.wait_for_transaction_receipt(tx_hash)