+++ eth_defi/token.py
@@ -12,7 +12,7 @@ from eth_tester.exceptions import TransactionFailed
 from eth_typing import HexAddress
 from web3 import Web3
 from web3.contract import Contract
-from web3.exceptions import BadFunctionCallOutput
+from web3.exceptions import BadFunctionCallOutput, ContractLogicError
 
 from eth_defi.abi import get_deployed_contract
 from eth_defi.deploy import deploy_contract
@@ -21,7 +21,7 @@ from eth_defi.utils import sanitise_string
 #: List of exceptions JSON-RPC provider can through when ERC-20 field look-up fails
 #: TODO: Add exceptios from real HTTPS/WSS providers
 #: `ValueError` is raised by Ganache
-_call_missing_exceptions = (TransactionFailed, BadFunctionCallOutput, ValueError)
+_call_missing_exceptions = (TransactionFailed, BadFunctionCallOutput, ValueError, ContractLogicError)
 
 
 @dataclass
@@ -81,6 +81,15 @@ class TokenDetails:
 
         """
         return int(decimal_amount * 10**self.decimals)
+
+    def fetch_balance_of(self, address: HexAddress | str) -> Decimal:
+        """Get an address token balance.
+
+        :return:
+            Converted to decimal using :py:meth:`convert_to_decimal`
+        """
+        raw_amount = self.contract.functions.balanceOf(address).call()
+        return self.convert_to_decimals(raw_amount)
 
 
 class TokenDetailError(Exception):
@@ -150,6 +159,9 @@ def fetch_erc20_details(
     :param raise_on_error: If set, raise `TokenDetailError` on any error instead of silently ignoring in and setting details to None.
     :return: Sanitised token info
     """
+
+    # No risk here, because we are not sending a transaction
+    token_address = Web3.to_checksum_address(token_address)
 
     erc_20 = get_deployed_contract(web3, "ERC20MockDecimals.json", token_address)